refactor: BaseEntity 추가하여 중복된 entity 필드 제거
