+++ src/main/java/com/codessquad/qna/question/Question.java
@@ -1,6 +1,7 @@
 package com.codessquad.qna.question;
 
 import com.codessquad.qna.answer.Answer;
+import com.codessquad.qna.common.BaseEntity;
 import com.codessquad.qna.exception.InsufficientAuthenticationException;
 import com.codessquad.qna.user.User;
 
@@ -9,11 +10,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 @Entity
-public class Question {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
+public class Question extends BaseEntity {
     @Column(nullable = false, length = 100)
     private String title;
 
@@ -21,8 +18,6 @@ public class Question {
     private String contents;
 
     private boolean deleted;
-    private LocalDateTime createDateTime;
-    private LocalDateTime updateDateTime;
 
     @ManyToOne
     @JoinColumn(foreignKey = @ForeignKey(name = "fk_question_writer"))
@@ -35,11 +30,10 @@ public class Question {
     }
 
     public Question(Long id, String title, String contents, LocalDateTime createDateTime, LocalDateTime updateDateTime, User writer, List<Answer> answers) {
-        this.id = id;
+        super(id, createDateTime, updateDateTime);
+
         this.title = title;
         this.contents = contents;
-        this.createDateTime = createDateTime == null ? LocalDateTime.now() : createDateTime;
-        this.updateDateTime = updateDateTime;
         this.writer = writer;
         this.answers = answers;
     }
@@ -97,10 +91,6 @@ public class Question {
         }
     }
 
-    public Long getId() {
-        return id;
-    }
-
     public String getTitle() {
         return title;
     }
@@ -113,28 +103,12 @@ public class Question {
         deleted = true;
     }
 
-    public LocalDateTime getCreateDateTime() {
-        return createDateTime;
-    }
-
-    public LocalDateTime getUpdateDateTime() {
-        return updateDateTime;
-    }
-
     public User getWriter() {
         return writer;
     }
 
-    public void setWriter(User writer) {
-        this.writer = writer;
-    }
-
     public List<Answer> getAnswers() {
         return answers;
-    }
-
-    public void setAnswers(List<Answer> answers) {
-        this.answers = answers;
     }
 
     public void verifyWriter(User target) {
@@ -159,17 +133,16 @@ public class Question {
 
         this.title = newQuestion.title;
         this.contents = newQuestion.contents;
-        this.updateDateTime = LocalDateTime.now();
     }
 
     @Override
     public String toString() {
         return "Question{" +
-                "id=" + id +
+                "id=" + getId() +
                 ", title='" + title + '\'' +
                 ", contents='" + contents + '\'' +
-                ", createDateTime=" + createDateTime +
-                ", updateDateTime=" + updateDateTime +
+                ", createDateTime=" + getCreateDateTime() +
+                ", updateDateTime=" + getUpdateDateTime() +
                 ", writer=" + writer +
                 ", answers=" + answers +
                 '}';
